Create DataLoaderCapacity
