+++ smartvolt/src/main/java/ua/tqs/smartvolt/smartvolt/data_loaders/capacity/DataLoaderCapacity.java
@@ -1,0 +1,100 @@
+package ua.tqs.smartvolt.smartvolt.data_loaders.capacity;
+
+import jakarta.transaction.Transactional;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import ua.tqs.smartvolt.smartvolt.models.EvDriver;
+import ua.tqs.smartvolt.smartvolt.models.StationOperator;
+import ua.tqs.smartvolt.smartvolt.repositories.BookingRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.ChargingSessionRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.ChargingSlotRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.ChargingStationRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.EvDriverRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.PaymentRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.StationOperatorRepository;
+import ua.tqs.smartvolt.smartvolt.repositories.UserRepository;
+
+@Component
+@Profile("capacity")
+public class DataLoaderCapacity implements CommandLineRunner {
+
+  private final PasswordEncoder passwordEncoder;
+  private final StationOperatorRepository stationOperatorRepository;
+  private final ChargingStationRepository chargingStationRepository;
+  private final ChargingSlotRepository chargingSlotRepository;
+  private final UserRepository userRepository;
+  private final EvDriverRepository evDriverRepository;
+  private final BookingRepository bookingRepository;
+  private final PaymentRepository paymentRepository;
+  private final ChargingSessionRepository chargingSessionRepository;
+
+  public DataLoaderCapacity(
+      StationOperatorRepository sor,
+      ChargingStationRepository csr,
+      ChargingSlotRepository cslr,
+      UserRepository ur,
+      EvDriverRepository evr,
+      BookingRepository bR,
+      PasswordEncoder passwordEncoder,
+      PaymentRepository paymentRepository,
+      ChargingSessionRepository chargingSessionRepository) {
+    this.stationOperatorRepository = sor;
+    this.chargingStationRepository = csr;
+    this.chargingSlotRepository = cslr;
+    this.userRepository = ur;
+    this.evDriverRepository = evr;
+    this.bookingRepository = bR;
+    this.passwordEncoder = passwordEncoder;
+    this.paymentRepository = paymentRepository;
+    this.chargingSessionRepository = chargingSessionRepository;
+  }
+
+  private void dropDatabase() {
+    // First, delete all payments and charging sessions
+    paymentRepository.deleteAll();
+    chargingSessionRepository.deleteAll();
+    // Then, delete all bookings referencing ChargingSlots
+    bookingRepository.deleteAll();
+
+    // Now, delete charging slots
+    chargingSlotRepository.deleteAll();
+
+    // Proceed with other deletions
+    chargingStationRepository.deleteAll();
+    stationOperatorRepository.deleteAll();
+    evDriverRepository.deleteAll();
+    userRepository.deleteAll();
+  }
+
+  @Override
+  @Transactional
+  public void run(String... args) throws Exception {
+    System.out.println("DataLoader is running...");
+
+    System.out.println("Test password");
+    System.out.println("Encoded password: " + passwordEncoder.encode("StrongPassword!"));
+
+    // Clear the database
+    dropDatabase();
+
+    StationOperator stationOperator =
+        new StationOperator(
+            "John Doe", "johndoe@example.com", passwordEncoder.encode("StrongPassword!"));
+    stationOperatorRepository.saveAndFlush(stationOperator);
+    System.out.printf(
+        "Station Operator created: %s with ID %s%n",
+        stationOperator.getName(), stationOperator.getUserId());
+
+    // Create a EVDriver
+    EvDriver testEVDriver =
+        new EvDriver("Jane Smith", "test@example.com", passwordEncoder.encode("password123!"));
+    evDriverRepository.saveAndFlush(testEVDriver);
+
+    System.out.printf(
+        "EV Driver created: %s with ID %s%n", testEVDriver.getName(), testEVDriver.getUserId());
+
+    System.out.println("DataLoader finished.");
+  }
+}
